+++ cleaned-code/timing.py
@@ -1,4 +1,4 @@
-from utils import *
+from util import *
 import time
 from datetime import datetime, timezone
